- Added pagination for all customers

+++ src/main/java/com/pekka/moni/customer/CustomerService.java
@@ -3,9 +3,10 @@ package com.pekka.moni.customer;
 import com.pekka.moni.exception.customer.CustomerAlreadyExistsException;
 import com.pekka.moni.exception.customer.CustomerNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 public class CustomerService {
@@ -23,8 +24,13 @@ public class CustomerService {
 
     }
 
-    public List<Customer> getCustomers() {
-        return customerRepository.findAll();
+    public Page<Customer> getCustomers(String sortBy, String sortDirection, int page, int pageSize) {
+        //TODO: Mapper to map Page<Customer> to Page<CustomerDto>
+        PageRequest pageRequest = PageRequest.of(page, pageSize);
+        return switch (Sort.Direction.fromString(sortDirection)) {
+            case ASC -> customerRepository.findAll(pageRequest.withSort(Sort.by(sortBy).ascending()));
+            case DESC -> customerRepository.findAll(pageRequest.withSort(Sort.by(sortBy).descending()));
+        };
     }
 
     public void addNewCustomer(Customer customer) {
@@ -38,14 +44,15 @@ public class CustomerService {
     }
 
     public void deleteCustomer(Long id) {
-        customerRepository.findById(id)
-                          .orElseThrow(() -> new CustomerNotFoundException("Customer with id " + id + " not found"));
+        if (!customerRepository.existsById(id)) {
+            throw new CustomerNotFoundException("Customer with id " + id + " not found");
+        }
         customerRepository.deleteById(id);
     }
 
     public void updateCustomer(Customer customer, Long id) {
         Customer customerToUpdate = customerRepository.findById(id)
-                                       .orElseThrow(() -> new CustomerNotFoundException("Customer with id " + id + " not found"));
+                                                      .orElseThrow(() -> new CustomerNotFoundException("Customer with id " + id + " not found"));
 
         customerToUpdate.setEmail(customer.getEmail());
         customerToUpdate.setFirstName(customer.getFirstName());
